Fixed issue with PUT

+++ src/main/java/com/qa/alysopedia/repository/WikiRepository.java
@@ -3,8 +3,8 @@ package com.qa.alysopedia.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.qa.alysopedia.model.Wiki;
+import com.qa.alysopedia.domain.Wiki;
 
 @Repository
-public interface WikiRepository extends MongoRepository<Wiki, Integer> {
+public interface WikiRepository extends MongoRepository<Wiki, Long> {
 }
